refactor: AddLinkField와 AddLinkBar 분리
