+++ components/AddLinkField/AddLinkField.module.scss
@@ -1,98 +1,40 @@
 @use "@/styles/abstract/variables_color" as var-color;
 @use "@/styles/abstract/mixins" as mixin;
 
-.formWrapper {
-  width: 100%;
-  max-width: 800px;
-  margin: 0 auto;
+.observedWrapper {
+  position: relative;
 
-  .form {
+  .addLinkContainer {
     width: 100%;
-    height: 69px;
-    background: var-color.$linkbrary-white;
-    border: 1px solid var-color.$linkbrary-primary;
-    border-radius: 15px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    .linkWrapper {
-      width: 100%;
-      height: 24px;
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-      background: transparent;
-      gap: 13px;
-      padding-left: 20px;
-
-      .linkIconWrapper {
-        position: relative;
-        width: 20px;
-        height: 20px;
-
-        .image {
-          @include mixin.image;
-        }
-      }
-
-      .inputWrapper {
-        width: 100%;
-        font-weight: 400;
-        font-size: 16px;
-        line-height: 24px;
-        text-align: left;
-        color: var-color.$linkbrary-gray-60;
-
-        .input {
-          margin: 0;
-          padding: 0;
-          width: 100%;
-          height: auto;
-          display: block;
-          border: none;
-          background: transparent;
-        }
-        .input:focus {
-          outline: none;
-        }
-      }
-    }
-
-    .button {
-      width: 80px;
-      min-width: 80px;
-      height: 37px;
-      margin-right: 20px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      background: linear-gradient(
-        90.99deg,
-        var-color.$linkbrary-primary 0.12%,
-        #6ae3fe 101.84%
-      );
-      border-radius: 8px;
-      color: #f5f5f5;
-      font-weight: 600;
-      font-size: 14px;
-      line-height: 17px;
-      display: flex;
-      align-items: center;
-      cursor: pointer;
-    }
+    padding: 60px 32px 90px;
+    background-color: var-color.$linkbrary-bg;
   }
-  // .observed {
-  //   height: 1px;
-  // }
+  .addLinkContainer.false {
+    position: fixed;
+    width: 100%;
+    background: var-color.$linkbrary-bg;
+    bottom: 0;
+    z-index: 99;
+    padding: 16px 32px 16px;
+  }
+
+  .observed {
+    height: 1px;
+    position: absolute;
+    bottom: 90px;
+  }
 }
 
 @include mixin.mobile {
-  .formWrapper {
-    max-width: calc(100% - 6.4rem);
-
-    .form {
-      height: 53px;
+  .observedWrapper {
+    .addLinkContainer {
+      padding: 24px 0 40px;
+    }
+    .addLinkContainer.false {
+      padding: 16px 0 16px;
+    }
+    .observed {
+      bottom: 0px;
     }
   }
 }
