+++ components/FolderContents/FolderContents.module.scss
@@ -6,36 +6,3 @@
   background: var-color.$linkbrary-white;
   margin: 0 auto;
 }
-.observedWrapper {
-  position: relative;
-}
-.observed {
-  height: 1px;
-  position: absolute;
-  bottom: 90px;
-}
-.addLinkContainer {
-  width: 100%;
-  padding: 60px 32px 90px;
-  background-color: var-color.$linkbrary-bg;
-}
-.addLinkContainer.false {
-  position: fixed;
-  width: 100%;
-  background: var-color.$linkbrary-bg;
-  bottom: 0;
-  z-index: 99;
-  padding: 16px 32px 16px;
-}
-
-@media only screen and (max-width: 768px) {
-  .addLinkContainer {
-    padding: 24px 0 40px;
-  }
-  .addLinkContainer.false {
-    padding: 16px 0 16px;
-  }
-  .observed {
-    bottom: 0px;
-  }
-}
